Remove Altitude Limit
Wind data was originally clamped to 11km altitude to match OpenRocket.  That limit was removed from OR as part of the release featuring multi-level wind.  It now provides the full altitude range provided by Open-Meteo.

Values in the CSV files are now limited to a single digit of precision to improve readability.

Also cleaned up the codebase.  This mostly involved deleting functionality inherited from GPS DC but not actually used.